+++ src/app/gigs/chat/chat.tsx
@@ -3,7 +3,7 @@ import { ChatList } from './chat-list';
 import ChatTopbar from './chat-topbar';
 import { useAuth } from '@/hooks';
 import { useChat } from '@/hooks/chat/useChat';
-import { ContractType, PrismaOfferType } from '@/types';
+import { useGetOffer } from '@/hooks/gigs/useGetOffer';
 import { File, ImageIcon } from 'lucide-react';
 import { useEffect, useState } from 'react';
 
@@ -17,39 +17,20 @@ type SupportedFileIcon = keyof typeof icons;
 
 interface ChatProps {
   showTopbar?: boolean;
-  contractData: ContractType | PrismaOfferType;
   isOpen: boolean;
+  currentOfferId: string;
+  contractStatus: string;
 }
 
-export function Chat({ showTopbar = true, contractData, isOpen }: ChatProps) {
-  const [isMobile, setIsMobile] = useState<boolean>(false);
+export function Chat({ showTopbar = true, isOpen, currentOfferId, contractStatus = '' }: ChatProps) {
   const { user, logout } = useAuth();
+  const { data: offerData } = useGetOffer(currentOfferId);
   const [state, chatLoading, error, sendMessage] = useChat({
     user,
     logout,
-    contractData,
     isOpen,
+    offerData,
   });
-
-  // This effect is mainly to disconnect from Sendbird so that when a new message is added, the connectionStatus property is
-  // properly being updated when a user closes out of the chat
-
-  useEffect(() => {
-    const checkScreenWidth = () => {
-      setIsMobile(window.innerWidth <= 768);
-    };
-
-    // Initial check
-    checkScreenWidth();
-
-    // Event listener for screen width changes
-    window.addEventListener('resize', checkScreenWidth);
-
-    // Cleanup the event listener on component unmount
-    return () => {
-      window.removeEventListener('resize', checkScreenWidth);
-    };
-  }, []);
 
   return (
     <div className='flex flex-col justify-between w-full h-full bg-gray-100 p-2 py-3 rounded-2xl '>
@@ -59,16 +40,15 @@ export function Chat({ showTopbar = true, contractData, isOpen }: ChatProps) {
         messages={state.messages}
         selectedUser={user!}
         typingMembers={state.typingMembers}
-        contractStatus={contractData.status}
         supportedIcons={icons}
         chatLoading={chatLoading}
       />
       <div className='mt-4'>
         <ChatBottombar
           sendMessage={sendMessage}
-          isMobile={isMobile}
+          isMobile={false}
           channel={state.channel}
-          contractStatus={contractData.status}
+          contractStatus={contractStatus}
           supportedIcons={icons}
         />
       </div>
