+++ src/hooks/gigs/useGetOffer.ts
@@ -1,17 +1,13 @@
 import { getOfferById } from '@/api/gigs';
-import { ContractType } from '@/types';
+import { ContractType, PrismaOfferType } from '@/types';
 import { useQuery } from '@tanstack/react-query';
 
-export const useGetOffer = (
-  offerId: string,
-  contractData: ContractType | undefined
-) => {
+export const useGetOffer = (offerId: string) => {
   return useQuery({
     queryKey: ['offer', offerId],
-    queryFn: async () => {
+    queryFn: async (): Promise<PrismaOfferType> => {
       const res = await getOfferById(offerId);
       return res.data;
     },
-    enabled: !contractData && offerId.length > 0,
   });
 };
